#236 deleting export related responce code added; more integration test added around export functionality

+++ django-nonrel/ocl/collection/views.py
@@ -417,7 +417,10 @@ class CollectionVersionExportView(ResourceAttributeChildMixin):
         version = self.get_object()
         if version.has_export():
             key = version.get_export_key()
-            key.delete()
+            if key:
+                key.delete()
+                return HttpResponse(status=200)
+
         return HttpResponse(status=204)
 
     def handle_export_collection_version(self):
